Moved liquid cooled motor and battery back to thermal, whoops
