+++ openconcept/propulsion/__init__.py
@@ -1,7 +1,7 @@
 # Components
 from .cfm56 import CFM56
 from .generator import SimpleGenerator
-from .motor import SimpleMotor, LiquidCooledMotor
+from .motor import SimpleMotor
 from .N3 import N3, N3Hybrid
 from .propeller import SimplePropeller, WeightCalc, ThrustCalc, PropCoefficients
 from .splitter import PowerSplit
