+++ openconcept/thermal/__init__.py
@@ -2,8 +2,8 @@ from .chiller import HeatPumpWithIntegratedCoolantLoop
 from .ducts import ExplicitIncompressibleDuct, ImplicitCompressibleDuct, ImplicitCompressibleDuct_ExternalHX
 from .heat_exchanger import HXGroup
 from .heat_pipe import HeatPipe
-from .battery_cooling import BandolierCoolingSystem
-from .motor_cooling import MotorCoolingJacket
+from .battery_cooling import BandolierCoolingSystem, LiquidCooledBattery
+from .motor_cooling import MotorCoolingJacket, LiquidCooledMotor
 from .pump import SimplePump
 from .hose import SimpleHose
 from .manifold import FlowSplit, FlowCombine
