+++ openconcept/thermal/battery_cooling.py
@@ -1,5 +1,110 @@
 import openmdao.api as om
 import numpy as np
+from openconcept.utilities import Integrator
+
+
+class LiquidCooledBattery(om.Group):
+    """A battery with liquid cooling 
+
+    Inputs
+    ------
+    q_in : float
+        Heat produced by the operating component (vector, W)
+    mdot_coolant : float
+        Coolant mass flow rate (vector, kg/s)
+    T_in : float
+        Instantaneous coolant inflow temperature (vector, K)
+    battery_weight : float
+        Battery weight (scalar, kg)
+    n_cpb : float
+        Number of cells long per "bandolier" actual count is 2x (scalar, default 82, Tesla)
+    t_channel : float
+        Thickness (width) of the cooling channel in the bandolier
+        (scalar, default 1mm)
+    T_initial : float
+        Initial temperature of the battery (only required in thermal mass mode) (scalar, K)
+    duration : float
+        Duration of mission segment, only required in unsteady mode
+
+    Outputs
+    -------
+    T_out : float
+        Instantaneous coolant outlet temperature (vector, K)
+    T: float
+        Battery volume averaged temperature (vector, K)
+    T_core : float
+        Battery core temperature (vector, K)
+    T_surface : float
+        Battery surface temperature (vector, K)
+        
+    Options
+    -------
+    num_nodes : int
+        Number of analysis points to run
+    quasi_steady : bool
+        Whether or not to treat the component as having thermal mass
+        num_nodes : float
+        The number of analysis points to run
+    coolant_specific_heat : float
+        Specific heat of the coolant (J/kg/K) (default 3801, glycol/water)
+    fluid_k : float
+        Thermal conductivity of the coolant (W/m/K)
+    nusselt : float
+        Nusselt number of the coolant channel (default 7.54 for uniform surf temp)
+    cell_kr : float
+        Thermal conductivity of the cell in the radial direction (W/m/k)
+    cell_diameter : float
+        Battery diameter (default 21mm for 2170 cell)
+    cell_height : float
+        Battery height (default 70mm for 2170 cell)
+    cell_mass : float
+        Battery weight (default 70g for 2170 cell)
+    cell_specific_heat : float
+        Mass average specific heat of the battery (default 900, LiIon cylindrical cell)
+    battery_weight_fraction : float
+        Fraction of battery by weight that is cells (default 0.72 knocks down Tesla by a bit)
+    """
+
+    def initialize(self):
+        self.options.declare('quasi_steady', default=False, desc='Treat the component as quasi-steady or with thermal mass')
+        self.options.declare('num_nodes', default=1, desc='Number of quasi-steady points to runs')
+        self.options.declare('coolant_specific_heat', default=3801, desc='Coolant specific heat in J/kg/K')
+        self.options.declare('fluid_k', default=0.405, desc='Thermal conductivity of the fluid in W / mK')
+        self.options.declare('nusselt', default=7.54, desc='Hydraulic diameter Nusselt number')
+
+        self.options.declare('cell_kr', default=0.3) # 0.455 for an 18650 cell, knocked down a bit
+        self.options.declare('cell_diameter', default=0.021)
+        self.options.declare('cell_height', default=0.070)
+        self.options.declare('cell_mass', default=0.070)
+        self.options.declare('cell_specific_heat', default=875.)
+        self.options.declare('battery_weight_fraction', default=0.65)
+    def setup(self):
+        nn = self.options['num_nodes']
+        quasi_steady = self.options['quasi_steady']
+        
+        self.add_subsystem('hex', BandolierCoolingSystem(num_nodes=nn, 
+                                                         coolant_specific_heat=self.options['coolant_specific_heat'],
+                                                         fluid_k=self.options['fluid_k'],
+                                                         nusselt=self.options['nusselt'],
+                                                         cell_kr=self.options['cell_kr'],
+                                                         cell_diameter=self.options['cell_diameter'],
+                                                         cell_height=self.options['cell_height'],
+                                                         cell_mass=self.options['cell_mass'],
+                                                         cell_specific_heat=self.options['cell_specific_heat'],
+                                                         battery_weight_fraction=self.options['battery_weight_fraction']),
+                            promotes_inputs=['q_in', 'mdot_coolant', 'T_in', ('T_battery', 'T'), 'battery_weight', 'n_cpb', 't_channel'],
+                            promotes_outputs=['T_core', 'T_surface', 'T_out', 'dTdt'])
+        
+        if not quasi_steady:
+            ode_integ = self.add_subsystem('ode_integ', Integrator(num_nodes=nn, diff_units='s', method='simpson', time_setup='duration'),
+                                           promotes_outputs=['*'], promotes_inputs=['*'])
+            ode_integ.add_integrand('T', rate_name='dTdt', units='K', lower=1e-10)
+        else:
+            self.add_subsystem('thermal_bal',
+                               om.BalanceComp('T', eq_units='K/s', lhs_name='dTdt', rhs_val=0.0, units='K', lower=1.0, val=299.*np.ones((nn,))),
+                               promotes_inputs=['dTdt'],
+                               promotes_outputs=['T'])
+
 
 class BandolierCoolingSystem(om.ExplicitComponent):
     """
